pyproject.toml
[project]
name = "conversor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.3.0",
    "customtkinter>=5.2.2",
    "flet>=0.28.3",
    "flet-desktop>=0.28.3",
    "pillow>=12.0.0",
    "rich>=14.2.0",
    "tkinterdnd2>=0.4.3",
]
